New Mexico State Legislature bills can be scraped now.
This is not a fully working implementation though. For one thing, sub sessions are not used when creating the bill object. More importantly, none of the documents listed for a bill are added to the bill object. Co-sponsors (if any) are not being added either. Finally, even though they list bills going back to the 1996 session, this script only scrapes bills from both 2010 sessions.

Lots of more work needs to be done.