WIP: New CO legislature site at http://leg.colorado.gov
This WIP commit adds the LegislatorScraper for the new site.  Because not all the terms published on the previous site (http://www.leg.state.co.us) are available on the new site, this does not delete or modify the previous scraper.  I've added a separate scraper for the new site, which is valid for terms "2015-2016" and up.

There's a little weirdness in that I actually have three Legislator scraper classes - one for the previous version of the site, one for the new, and one that does nothing except choose between these two scrapers based on the term passed to `scrape`.  There are some internal attributes updated on each scraper instance, which I manually copied in a hacky way to the "chooser" scraper.

TODOs:
These scrapers still need to be updated for the new site.

- [ ] Bills
- [ ] Committees
- [ ] Actions
- [ ] Votes