NJ: Save all committees after members have been added
`save_committee` was only being called once, with the last committee to have a member added to it.  Because of this, only one committee was ever saved per session.

This diff updates the script to iterate over the scraped committees and save each one at the end of the script.