Update parser for committee chair name to account for no FULLNAME
Instead, parse the KIPD provided in a bit of a rudementary way, assuming
a standard format for the ID is consistent. In this case, seems like the
committee chair was NOT a legislator and so the full information was not
provided. However, info could be gotten out of a (hopefully)
  consistently formed ID.