MO: Make "Assigned bills" filter for committee links case insensitive
In addition to "Assigned bills", the link text was also sometimes "Assigned Bills".  This was allowing some links to committee bills to be used as the committee details page, which caused the script to crash.

This also skips sessions without any session details from being scraped.  The script was crashing when it tried to process the 2018 session of the "2017-2018" term, which is not available yet.