NC: Strip leading "Rep. " and "Sen. " from committee members to help with matching
This strips leading "Rep. " and "Sen. " from committee members to help with name matching. NameMatcher seems like it is treating "Rep. Adams", for example, as a person whose first name is "Rep. ".

This silences the warning "No matches for" for all but two members, "Brenden Jones", who is listed as "Brendan H. Jones" on the legislators page, and "R.Turner".