Restructured to efficiently retreive photo_url
This restructure the RI to efficiently retreive the photo_url.

This uses the contact information page to retreive the detail link, instead of trying to match
the link to the representatives name.  On the contact information page, the link is associated
directly to the district.

This moves retreiving the contact information outside of the spreadsheet loop so we can utilize
the contact information more efficiently.

This removes the representatives listings page, e.g. http://www.rilin.state.ri.us/senators/default.aspx, as a source, since it is no longer used.

This prefers the email address in the contact information to that in the spreadsheet. I manually verified that the contact information always has valid email addresses, while the spreadhsheet sometimes has web page links.

This fixes the type "Dictrict Office" to "District Office".

This runs about twice as fast locally.

Closes #1470 and prevents future typos for #1474