Migrated Event scraper
This involved some strangeness with timezones. I am saving the datetime _incorrectly_: I localize the time listed on the page (which is central time) to UTC, but include the correct timezone in the Event timezone attribute.  Doing this, the time is saved 5 hours in advance, and the datetime object passes validation.