TN: Fix vote scraping to pull from "Votes" section
This fixes the TN scraper to pull votes from the bill page "Votes" section. See examples in #1559.

This contains a small refactor that splits out the function for scraping each house's votes. This actually minimizes whitespace changes in the diff.  Instead of adding votes within the vote scrape method, it returns a list of votes which each are added to the bill.