Split lines and handle extra columns for NH roll call votes
Fixes two issues with the NH voter name scrape. Similar to an earlier fix for the vote counts, this adds `splitlines()` to the file so that it is processed line by line instead of character by character. In addition, this handles an extra two columns being returned when splitting each line.

Confirmed that Votes were not being saved with the bug in the voter name scrapee, and they are with the fix. This fixes the underlying cause of #1563.