Update Puerto Rico people scrapers
Fixes #1611. This is a pretty complete overhaul of the people scraper for Puerto Rico, as both the House and Senate pages changed. We now have to scrape the page for each individual legislator to get the information we want, as well as occasionally getting information from the main legislators page as some information is easier to scrape from that page.