MD: Don't include phone numbers in address
I accidentially forgot to use an elif in #2460, so when a phone number was included in the address field, it would be parsed and included in the parsed address as well. Clean it up by only adding it to the parsed address if it's not a fax line or a phone line.