IN: Fix people and committee scrapers
This includes many fixes. First, don't limit the session name to 5 characters as this cuts off the end of `2018ss1`. Secondly, `senate` and `house` apparently need to be completely lowercase for the API URLs. Thirdly, adjust the read timeout to start at 8 seconds, as 2 seconds seems to be too low for many pages.

Finally, add `2018ss1` as an acceptable session for the 2017-2018 term, and fix a typo in the legislative term to be `2017-2018` instead of `2017-2017`.

Fixes #2228.