MI: avoid error when vote roll call data is not as expected
Scrape was halting at an error NoneType object is not subscriptable when parse_roll_call() returned None. Check for None before calling the vote yielding logic. Also fixed the guard in parse_roll_call() so that it can find the text its looking for even when there are spaces/newlines/span tags between the words (which was causing None to be returned).