IN: Switch title to bill description
Currently, the API's bill title is used as its title to store in the database. However, this title is not very helpful. For instance, this year's HB 1001 title is:

A BILL FOR AN ACT concerning education.

Versus the description:

School accountability.

This changes the stored title to be the bill description and not the API title.