PA: Strip vote from member name in roll call votes
In roll call votes, the bill scraper grabs all the text in the vote div as the member's name. However, the div also contains a Y/N/E/X in a child element to record what the member's vote was. By using only the text of the div itself, we will only capture the member's name and not their vote.