MI: Convert newlines in roll call votes to spaces
The House and Senate journals occasionally have newlines in place of spaces, leading to the phrases "Not Voting" and "In The Chair" not being recognized. Replace these newline characters with spaces so the phrases are recognized correctly.

Also split lines with names on multiple spaces as Rep. O'Malley's name is sometimes recorded in the journal as "O Malley", causing his name to be split across two votes.