VA: Fix vote scraping
Vote scraping in VA has been broken, as the method erased any `cached_vote`s after a continue line was commented out. In order to remove much of the complexity around processing each line independently, we now process actions along with the following one. If it appears the following action is related to the current one, we merge the details together like in the previous method.