MO, NH, NJ, NM, SC, UT: Clean legislator names
Scrapers in these states were often not able to match legislators since they were not formatted as expected. In every state except for NH, the names were formatted as last, first. This change flips the formatting for names in these states to be first, last. In NH, additional spaces were accidentally added to names, which this change prevents.